+++ packages/dev/src/entry-server.jsx
@@ -3,4 +3,6 @@ import shopifyConfig from '../shopify.config';
 
 import App from './App.server';
 
-export default renderHydrogen(App, {shopifyConfig});
+const pages = import.meta.globEager('./pages/**/*.server.[jt](s|sx)');
+
+export default renderHydrogen(App, {shopifyConfig, pages});