+++ packages/hydrogen/src/foundation/Router/tests/DefaultRoutes.test.tsx
@@ -1,6 +1,8 @@
-import {createRoutesFromPages, ImportGlobEagerOutput} from '../DefaultRoutes';
+import {createRoutesFromPages} from '../DefaultRoutes';
+import {ImportGlobEagerOutput} from '../../../types';
+import {Logger} from '../../..';
 
-const STUB_MODULE = {default: {}};
+const STUB_MODULE = {default: {}, api: null};
 
 it('converts normal pages to routes', () => {
   const pages: ImportGlobEagerOutput = {
@@ -240,4 +242,17 @@ it('factors in the top-level path prefix', () => {
       exact: true,
     },
   ]);
+});
+
+it("errors routes don't have a default or api export", () => {
+  const log: Logger = {...console, fatal: jest.fn(), warn: jest.fn()};
+  const pages: ImportGlobEagerOutput = {
+    './pages/contact.server.jsx': {} as any,
+  };
+
+  createRoutesFromPages(pages, '*', log);
+
+  expect(log.warn).toBeCalledWith(
+    `./pages/contact.server.jsx doesn't export a default React component or an API function`
+  );
 });