+++ packages/hydrogen/src/framework/Hydration/ServerComponentRequest.server.ts
@@ -32,6 +32,10 @@ export class ServerComponentRequest extends Request {
       super(getUrlFromNodeRequest(input), {
         headers: new Headers(input.headers as {[key: string]: string}),
         method: input.method,
+        body:
+          input.method !== 'GET' && input.method !== 'HEAD'
+            ? input.body
+            : undefined,
       });
     }