+++ packages/hydrogen/src/types.ts
@@ -1,6 +1,8 @@
 import {ServerResponse} from 'http';
 import type {ServerComponentRequest} from './framework/Hydration/ServerComponentRequest.server';
 import type {Metafield, Image, MediaContentType} from './graphql/types/types';
+import {ApiRouteMatch} from './utilities/apiRoutes';
+import {Logger} from './utilities/log/log';
 
 export type Renderer = (
   url: URL,
@@ -37,6 +39,8 @@ export type EntryServerHandler = {
   render: Renderer;
   stream: Streamer;
   hydrate: Hydrator;
+  getApiRoute: (url: URL) => ApiRouteMatch | null;
+  log: Logger;
 };
 
 export type ShopifyConfig = {
@@ -50,16 +54,20 @@ export type Hook = (
   params: {url: URL} & Record<string, any>
 ) => any | Promise<any>;
 
+export type ImportGlobEagerOutput = Record<
+  string,
+  Record<'default' | 'api', any>
+>;
+
+export type ServerHandlerConfig = {
+  pages?: ImportGlobEagerOutput;
+  shopifyConfig: ShopifyConfig;
+};
+
 export type ServerHandler = (
   App: any,
-  options: {
-    shopifyConfig: ShopifyConfig;
-  }
-) => {
-  render: Renderer;
-  stream: Streamer;
-  hydrate: Hydrator;
-};
+  config?: ServerHandlerConfig
+) => EntryServerHandler;
 
 export type ClientHandler = (
   App: any,