+++ packages/playground/server-components-worker/src/App.server.jsx
@@ -2,12 +2,10 @@ import {DefaultRoutes} from '@shopify/hydrogen';
 import {Suspense} from 'react';
 
 export default function App({...serverState}) {
-  const pages = import.meta.globEager('./pages/**/*.server.[jt]sx');
-
   return (
     <Suspense fallback={'Loading...'}>
       <DefaultRoutes
-        pages={pages}
+        pages={serverState.pages}
         serverState={serverState}
         fallback="Not Found"
       />